added error and case handling to gcards
