+++ src/new_script_generator.py
@@ -4,38 +4,13 @@ from script_generators import startup,initialization,run_gemc,run_evio2hipo,run_
 from utils import utils, file_struct, scard_helper, user_validation, gcard_helper
 from condor_scripts import condor_startup, condor_1, condor_2
 
-def grabber(scard_file):
-  funcs = (startup,initialization,run_gemc,run_evio2hipo,run_cooking,file_mover)
-  fname = ('startup','initialization','run_gemc','run_evio2hipo','run_cooking','file_mover')
-  funcs_condor = (condor_startup,condor_1,condor_2)
-  fname_condor = ('condor_startup','condor_1','condor_2')
-
-
-  BatchID = 1
-  strn = "SELECT scard FROM Batches WHERE BatchID = {};".format(BatchID)
-  scard_text = utils.sql3_grab(strn)[0][0]
-#  print(scard_text)
-  scard = scard_helper.scard_class(scard_text)
-  username = user_validation.user_validation()
-
-  #Write scard into scard table fields (This will not be needed in the future)
-  print("\nReading in information from {0}".format(scard_file))
-  utils.printer("Writing SCard to Database")
-  scard.data['genExecutable'] = file_struct.genExecutable.get(scard.data.get('generator'))
-  scard.data['genOutput'] = file_struct.genOutput.get(scard.data.get('generator'))
-  #scard_helper.SCard_Entry(BatchID,timestamp,scard_fields.data)
-  #print('\t Your scard has been read into the database with BatchID = {0} at {1} \n'.format(BatchID,timestamp))
-
-  for item in scard.data:
-    #print(scard.data.get(item))
-    print(item)
+def script_generator(script_name,gen_funcs,func_names,scard):
   newfile = "runscript.sh"
   if os.path.isfile(newfile):
     subprocess.call(['rm',newfile])
   for count, f in enumerate(funcs):
     generated_text = getattr(f,fname[count])(scard)
     with open(newfile,"a") as file: file.write(generated_text)
-
 
   newfile = "clas12.condor"
   if os.path.isfile(newfile):
@@ -44,24 +19,30 @@ def grabber(scard_file):
     generated_text = getattr(f,fname_condor[count])(scard)
     with open(newfile,"a") as file: file.write(generated_text)
 
+def grabber(scard_file,BatchID):
+  funcs = (startup,initialization,run_gemc,run_evio2hipo,run_cooking,file_mover)
+  fname = ('startup','initialization','run_gemc','run_evio2hipo','run_cooking','file_mover')
+  funcs_condor = (condor_startup,condor_1,condor_2)
+  fname_condor = ('condor_startup','condor_1','condor_2')
 
+  strn = "SELECT scard FROM Batches WHERE BatchID = {};".format(BatchID)
+  scard_text = utils.sql3_grab(strn)[0][0] #sql3_grab returns a list of tuples, we need the 0th element of the 0th element
+  scard = scard_helper.scard_class(scard_text)
+  username = user_validation.user_validation()
 
-"""
-    with open(scard_file, 'r') as file: scard = file.read()
-    strn = "UPDATE Batches SET {0} = '{1}' WHERE BatchID = "{2}";"""""".format('scard',scard,BatchID)
-    utils.sql3_exec(strn)
-"""
-"""
+  print("\nReading in information from {0}".format(scard_file))
+  utils.printer("Writing SCard to Database")
+  scard.data['genExecutable'] = file_struct.genExecutable.get(scard.data.get('generator'))
+  scard.data['genOutput'] = file_struct.genOutput.get(scard.data.get('generator'))
 
-def grab_gcards(BatchID):
-  strn = "SELECT GcardID, gcard_text FROM GCards WHERE BatchID = {};".format(BatchID)
-  gcards = utils.sql3_grab(strn)
-  return gcards
+  """
+  Need to have block of code for grabbing appropriate gcard. For now, just assume we grab
+  the standard gcard so we don't have to worry about this
+  """
 
+  script_generator("runscript.sh",funcs,fname,scard):
+  script_generator("clas12.condor",funcs_condor,fname_condor,scard):
 
-
-
-"""
 if __name__ == "__main__":
   argparser = argparse.ArgumentParser()
   argparser.add_argument('-s','--scard', default=file_struct.scard_path+file_struct.scard_name,
@@ -72,4 +53,5 @@ if __name__ == "__main__":
 
   file_struct.DEBUG = getattr(args,file_struct.debug_long)
   scard_file = args.scard
-  grabber(scard_file)
+  BatchID = 1
+  grabber(scard_file,BatchID)
